+++ core/ui.js
@@ -67,6 +67,7 @@ CKEDITOR.ui.prototype = {
 	 * Retrieves the created UI objects by name.
 	 *
 	 * @param {String} name The name of the UI definition.
+	 * @returns {Object} The UI object.
 	 */
 	get: function( name ) {
 		return this.instances[ name ];